# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auth-app API",
    "description": "The JSON API the admin frontend consumes.",
    "contact": {
      "name": "auth-app template"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Handle `POST /auth/login`.",
        "description": "`200` with the admin id and fresh session/CSRF cookies on success; `401`\nfor invalid credentials (identical for a wrong password and a nonexistent\naccount); `429` with `Retry-After` when the IP or account has hit the\napp-level rate limit, or when the account/IP is locked out (see\n[`LoginError::TooManyAttempts`] — a distinct mechanism from the rate limit:\nlockout counts *failures*, the rate limit counts *every* attempt); `500` on\nan internal failure. Every successful login issues a brand-new session —\nthere is no \"reuse the prior session\" path — which is what satisfies\nsession rotation on login.",
        "operationId": "login_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in; sets session and csrf cookies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOk"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials (wrong password or unknown account)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request (per-field validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate-limited or locked out; carries Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "The JSON body every error response carries.",
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable, machine-readable error code (e.g. `unauthorized`,\n`not_found`, `validation_failed`)."
          },
          "fields": {
            "type": [
              "object",
              "null"
            ],
            "description": "Per-field validation messages, present only on a `422` response.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "message": {
            "type": "string",
            "description": "A human-readable, client-safe message. Never contains internal detail."
          },
          "trace_id": {
            "type": "string",
            "description": "A correlation id, also logged server-side, so a client-reported error\ncan be matched to its log line."
          }
        }
      },
...
      "LoginBody": {
        "type": "object",
        "description": "The JSON body of a login request.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "The administrator's email address.",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "description": "The administrator's password."
          }
        }
      },
      "LoginOk": {
        "type": "object",
        "description": "The success body: the authenticated administrator's id.",
        "required": [
          "admin_id"
        ],
        "properties": {
          "admin_id": {
            "type": "string",
            "description": "The authenticated administrator's opaque id."
          }
        }
      },
...
    }
  },
...
}